modules/sagemaker/sagemaker-templates-service-catalog/templates/model_deploy/seed_code/deploy_app/requirements.txt
aws-cdk-lib==2.128.0
constructs>=10.0.0,<11.0.0
yamldataclassconfig